adding ability to sign actions
