+++ src/lib/components/sandbox/arena/ArenaCanvas.svelte
@@ -1,30 +1,17 @@
 <script lang="ts">
-  import ArenaCanvasMovementPhase from "./ArenaCanvasMovementPhase.svelte";
-  import ArenaCanvasShootingPhase from "./ArenaCanvasShootingPhase.svelte";
-  import ArenaCanvasMeleePhase from "./ArenaCanvasMeleePhase.svelte";
+	import ArenaCanvasMovementPhase from './arena-phases/ArenaCanvasMovementPhase.svelte';
+	import ArenaCanvasShootingPhase from './arena-phases/ArenaCanvasShootingPhase.svelte';
+	import ArenaCanvasMeleePhase from './arena-phases/ArenaCanvasMeleePhase.svelte';
 
-  export let game: Game;
-  export let playerColors: Array<string>;
-  export let rerender: () => {};
-
+	export let game: Game;
+	export let playerColors: Array<string>;
+	export let rerender: () => {};
 </script>
 
 {#if game.currentPhase?.name === 'MOVEMENT'}
-  <ArenaCanvasMovementPhase
-    game={game}
-    playerColors={playerColors}
-    rerender={rerender}
-  />
+	<ArenaCanvasMovementPhase {game} {playerColors} {rerender} />
 {:else if game.currentPhase?.name === 'SHOOTING'}
-  <ArenaCanvasShootingPhase
-    game={game}
-    playerColors={playerColors}
-    rerender={rerender}
-  />
+	<ArenaCanvasShootingPhase {game} {playerColors} {rerender} />
 {:else if game.currentPhase?.name === 'MELEE'}
-  <ArenaCanvasMeleePhase
-    game={game}
-    playerColors={playerColors}
-    rerender={rerender}
-  />
+	<ArenaCanvasMeleePhase {game} {playerColors} {rerender} />
 {/if}
