+++ src/lib/components/sandbox/arena/arena-phases/ArenaCanvasMeleePhase.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
 	import { afterUpdate, onMount } from 'svelte';
-	import * as Utils from '../play/utils';
+	import * as Utils from '../../play/utils';
 	import { DiceRollServiceClient } from '$lib/dice-service-client/DiceRollServiceClient';
 	import { MinaArenaClient } from '$lib/mina-arena-graphql-client/MinaArenaClient';
-	import HoveredGamePieceTooltipShooting from './HoveredGamePieceTooltipShooting.svelte';
-	import SubmitPhaseButton from './SubmitPhaseButton.svelte';
+	import HoveredGamePieceTooltipMelee from '../tooltip/HoveredGamePieceTooltipMelee.svelte';
+	import SubmitPhaseButton from '../SubmitPhaseButton.svelte';
+	import { player1, dummyPlayer } from '$lib/stores/sandbox/playerStore';
 
 	export let game: Game;
 	export let playerColors: Array<string>;
@@ -28,7 +29,7 @@
 	let orders: Record<string, Array<GamePieceOrder>>;
 	let issuingOrder: GamePieceOrder | undefined;
 
-	const SHOOTING_ARROW_COLOR = 'darkred';
+	const MELEE_ARROW_COLOR = 'darkred';
 
 	const minaArenaClient = new MinaArenaClient();
 	const playerPublicKeys = (game.gamePlayers || []).map((p) => p.player.minaPublicKey) || ['', ''];
@@ -58,67 +59,63 @@
 	afterUpdate(() => {
 		Utils.clearCanvas(ctx);
 		Utils.drawArenaBackground(ctx);
-		drawShootingPhaseUnderPieces(canvas, orders, selectedPiece);
 		drawPieces();
-		drawShootOrders(ctx, orders);
+		drawMeleePhase(canvas, orders, selectedPiece);
 	});
 
 	const drawPieces = () => {
 		Utils.drawAllPieces(canvas, ctx, drawnPieces, hoveredPiece, selectedPiece);
 	};
 
-	export const drawShootingPhaseUnderPieces = (
+	export const drawMeleePhase = (
 		canvas: HTMLCanvasElement,
 		orders: Record<string, Array<GamePieceOrder>>,
 		selectedPiece?: GamePiece
 	) => {
 		const ctx = canvas.getContext('2d')!;
 		if (selectedPiece) {
-			const unit = selectedPiece.playerUnit.unit;
-			const missileRange = unit.rangedRange || 0;
-
-			if (missileRange) {
-				Utils.drawMissileRangeCircle(ctx, orders, selectedPiece);
-			}
+			Utils.drawMeleeRangeCircle(ctx, orders, selectedPiece);
 		}
+
+		drawMeleeOrders(ctx, orders);
 	};
 
-	const drawShootOrders = (
+	const drawMeleeOrders = (
 		ctx: CanvasRenderingContext2D,
 		orders: Record<string, Array<GamePieceOrder>>
 	) => {
 		Object.keys(orders).forEach((gamePieceId) => {
 			orders[gamePieceId].forEach((gamePieceOrder: GamePieceOrder) => {
-				const rangedAttackAction = gamePieceOrder.rangedAttack;
-				if (rangedAttackAction) drawRangedAttackArrow(ctx, rangedAttackAction);
+				const meleeAttackAction = gamePieceOrder.meleeAttack;
+				if (meleeAttackAction) drawMeleeAttackArrow(ctx, meleeAttackAction);
 			});
 		});
 	};
 
-	const drawRangedAttackArrow = (
+	const drawMeleeAttackArrow = (
 		ctx: CanvasRenderingContext2D,
-		rangedAttackAction: RangedAttackAction
+		meleeAttackAction: MeleeAttackAction
 	) => {
-		const shootingPieceId = rangedAttackAction.gamePieceId;
-		const targetPieceId = rangedAttackAction.action.targetGamePieceId;
-		const shootingPiece = Utils.gamePieceById(shootingPieceId, gamePieces);
+		const attackingPieceId = meleeAttackAction.gamePieceId;
+		const targetPieceId = meleeAttackAction.action.targetGamePieceId;
+		const attackingPiece = Utils.gamePieceById(attackingPieceId, gamePieces);
 		const targetPiece = Utils.gamePieceById(targetPieceId, gamePieces);
 
-		if (!shootingPiece) {
-			console.log(`ERROR drawing shooting orders, no GamePiece found with ID ${shootingPieceId}`);
+		if (!attackingPiece) {
+			console.log(`ERROR drawing melee orders, no GamePiece found with ID ${attackingPieceId}`);
 			return;
 		}
 
 		if (!targetPiece) {
-			console.log(`ERROR drawing shooting orders, no GamePiece found with ID ${targetPieceId}`);
+			console.log(`ERROR drawing melee orders, no GamePiece found with ID ${targetPieceId}`);
 			return;
 		}
 
 		Utils.drawArrowWithOffset(
 			ctx,
-			shootingPiece.coordinates,
+			attackingPiece.coordinates,
 			targetPiece.coordinates,
-			SHOOTING_ARROW_COLOR,
+			MELEE_ARROW_COLOR,
 			Utils.PIECE_RADIUS - 4
 		);
 	};
@@ -129,7 +126,7 @@
 		hoveredPiece = Utils.pieceAtCanvasPoint(mouseCanvasPoint, drawnPieces, gamePieces);
 		if (hoveredPiece) {
 			onGamePieceHovered(hoveredPiece, mouseAbsolutePoint);
-			if (issuingOrder) draftShootOrder(hoveredPiece);
+			if (issuingOrder) draftMeleeOrder(hoveredPiece);
 		} else {
 			onGamePieceUnhovered();
 		}
@@ -138,20 +135,20 @@
 	const onMouseDown = (e: MouseEvent) => {
 		const mouseCanvasPoint = Utils.getMouseCanvasPoint(e, canvas);
 		const clickedPiece = Utils.pieceAtCanvasPoint(mouseCanvasPoint, drawnPieces, gamePieces);
-		if (selectedPiece && clickedPiece) draftShootOrder(clickedPiece);
+		if (selectedPiece && clickedPiece) draftMeleeOrder(clickedPiece);
 	};
 
 	const onMouseUp = (e: MouseEvent) => {
 		const mouseCanvasPoint = Utils.getMouseCanvasPoint(e, canvas);
 		const clickedPiece = Utils.pieceAtCanvasPoint(mouseCanvasPoint, drawnPieces, gamePieces);
 		if (issuingOrder && clickedPiece) {
-			finalizeShootOrder();
+			finalizeMeleeOrder();
 		} else {
 			selectedPiece = Utils.pieceAtCanvasPoint(mouseCanvasPoint, drawnPieces, gamePieces);
 		}
 	};
 
-	const draftShootOrder = (targetPiece: GamePiece) => {
+	const draftMeleeOrder = (targetPiece: GamePiece) => {
 		if (!selectedPiece) return;
 
 		// Validate that the selected piece is owned by the
@@ -165,15 +162,12 @@
 			return;
 		}
 
-		// Validate that the selected piece can make ranged attacks and is in range
-		const selectedUnit = selectedPiece.playerUnit.unit;
-		if (!selectedUnit.rangedNumAttacks || !selectedUnit.rangedRange) return;
-
-		const shootDistance = Utils.distanceBetweenPoints(
+		// Validate that the selected piece is in range
+		const attackDistance = Utils.distanceBetweenPoints(
 			selectedPiece.coordinates,
 			targetPiece.coordinates
 		);
-		if (shootDistance > selectedUnit.rangedRange) return;
+		if (attackDistance > Utils.MELEE_ATTACK_RANGE) return;
 
 		const placeholderDiceRoll = {
 			publicKey: {
@@ -187,30 +181,27 @@
 			}
 		};
 		issuingOrder = {
-			rangedAttack: {
+			meleeAttack: {
 				gamePieceId: selectedPiece.id,
 				action: {
 					targetGamePieceId: targetPiece.id,
-					diceRolls: placeholderDiceRoll
+					diceRolls: placeholderDiceRoll,
+					gamePieceNumber: selectedPiece.gamePieceNumber,
+					targetGamePieceNumber: targetPiece.gamePieceNumber
 				}
 			}
 		};
 		orders[selectedPiece.id] = [issuingOrder];
 	};
 
-	const finalizeShootOrder = () => {
-		if (
-			!selectedPiece ||
-			!issuingOrder ||
-			!issuingOrder.rangedAttack ||
-			!selectedPiece.playerUnit.unit.rangedRange
-		) {
+	const finalizeMeleeOrder = () => {
+		if (!selectedPiece || !issuingOrder || !issuingOrder.meleeAttack) {
 			issuingOrder = undefined;
 			selectedPiece = undefined;
 			return;
 		}
 
-		const targetPieceId = issuingOrder.rangedAttack.action.targetGamePieceId;
+		const targetPieceId = issuingOrder.meleeAttack.action.targetGamePieceId;
 		const targetPiece = Utils.gamePieceById(targetPieceId, gamePieces);
 
 		if (!targetPiece || targetPiece.gamePlayer.player.minaPublicKey === currentPlayerMinaPubKey) {
@@ -219,11 +210,11 @@
 			return;
 		}
 
-		const shootDistance = Utils.distanceBetweenPoints(
+		const attackDistance = Utils.distanceBetweenPoints(
 			selectedPiece.coordinates,
 			targetPiece.coordinates
 		);
-		if (shootDistance <= selectedPiece.playerUnit.unit.rangedRange) {
+		if (attackDistance <= Utils.MELEE_ATTACK_RANGE) {
 			orders[selectedPiece.id] = [issuingOrder];
 		} else {
 			orders[selectedPiece.id] = [];
@@ -233,39 +224,40 @@
 	};
 
 	const submitPhase = async () => {
-		const rangedAttackActions: Array<RangedAttackAction> = [];
-
-		isLoading = true;
+		const meleeAttackActions: Array<MeleeAttackAction> = [];
 		await Promise.all(
 			Object.values(orders)
 				.flat()
-				.map(async (shootOrder: GamePieceOrder) => {
-					if (!shootOrder.rangedAttack) return;
-					const attackingPieceId = shootOrder.rangedAttack.gamePieceId;
+				.map(async (meleeOrder: GamePieceOrder) => {
+					if (!meleeOrder.meleeAttack) return;
+					const attackingPieceId = meleeOrder.meleeAttack.gamePieceId;
 					const attackingPiece = Utils.gamePieceById(attackingPieceId, gamePieces);
 					const attackingUnit = attackingPiece?.playerUnit.unit;
-					const numAttacks = attackingUnit?.rangedNumAttacks || 1;
+					const numAttacks = attackingUnit?.meleeNumAttacks || 1;
 					const diceRolls: DiceRollInput[] = [];
 					console.log(
 						`Rolling for ${numAttacks} attacks from Piece ${attackingPiece?.playerUnit.name} (ID: ${attackingPieceId})`
 					);
 					for (let i = 0; i < numAttacks; i++) {
 						const order = {
-							...shootOrder
+							...meleeOrder
 						};
-						if (!order.rangedAttack) return;
+						if (!order.meleeAttack) return;
 						const diceRoll = await rollDiceForAttack();
 						diceRolls.push(diceRoll);
-						order.rangedAttack.action.diceRolls = diceRoll;
-						rangedAttackActions.push(order.rangedAttack);
+						order.meleeAttack.action.diceRolls = diceRoll;
+						meleeAttackActions.push(order.meleeAttack);
 					}
 				})
 		);
-		await minaArenaClient.submitShootingPhase(
-			currentPlayerMinaPubKey,
+		isLoading = true;
+		const player = currentPlayerMinaPubKey === $player1.publicKey ? $player1 : $dummyPlayer;
+		await minaArenaClient.submitMeleePhase(
+			player.publicKey,
 			game.id,
 			game.currentPhase!.id,
-			rangedAttackActions
+			meleeAttackActions,
+			player.privateKey
 		);
 		await rerender();
 	};
@@ -311,7 +303,7 @@
 		</div>
 	</tr>
 </table>
-<HoveredGamePieceTooltipShooting
+<HoveredGamePieceTooltipMelee
 	{game}
 	{hoveredPiece}
 	{selectedPiece}
