+++ src/lib/mina-arena-graphql-client/queries/get-game.ts
@@ -53,6 +53,7 @@ export const GetGameQuery = gql`
           x
           y
         }
+        gamePieceNumber
         health
       }
       arena {
