+++ src/routes/barracks/+page.svelte
@@ -1,9 +1,9 @@
 <script>
 	import DraftNewUnits from '$lib/components/sandbox/squad-selection/DraftNewUnits.svelte';
 	import MyUnits from '$lib/components/sandbox/squad-selection/MyUnits.svelte';
-	import { dummyPlayer1 } from '$lib/stores/sandbox/playerStore';
+	import { player1 } from '$lib/stores/sandbox/playerStore';
 
-	const player = dummyPlayer1;
+	const player = $player1.publicKey;
 </script>
 
 <div>
